tests: Move common citrea code to its own file.
